yolo_model_1/requirements.txt
fastapi
uvicorn[standard]
ultralytics
python-multipart
opencv-python-headless
Pillow